pyproject.toml
[project]
name = "aitana"
version = "0.1.0"
description = "AItana – AI-powered Telegram agent"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=22.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
